packages/model/requirements.txt
opencv-python==4.10.0.84
numpy==1.23.5
requests==2.28.1
python-dotenv==1.0.0
bleak==0.21.1
firebase-admin==6.4.0
Flask==2.3.2
Flask-Cors==4.0.1